removed tests and improved fetch.ts implementation
