+++ frontend/package.json
@@ -31,7 +31,6 @@
   "devDependencies": {
     "@eslint/js": "^9.20.0",
     "@types/material-ui": "^0.21.18",
-    "@types/node": "^22.13.4",
     "@types/react": "^19.0.8",
     "@types/react-dom": "^19.0.3",
     "@vitejs/plugin-react": "^4.3.4",
