+++ frontend/src/constants/config.ts
@@ -1,0 +1,1 @@
+export const BASE_URL = 'http://uw-floor-it.web.app/';
