--- frontend/tsconfig.script.json
@@ -1,15 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "ES2015",
-    "module": "ES2020",
-    "lib": ["ES2015", "DOM"],
-    "moduleResolution": "node",
-    "esModuleInterop": true,
-    "strict": true,
-    "outDir": "dist",
-    "skipLibCheck": true,
-    "types": ["node"]
-  },
-  "include": ["src/scripts/print_champion_titles.ts", "src/utils/fetch.ts"],
-  "exclude": ["node_modules"]
-}
